Implement Authorication using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -67,7 +67,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403,'Access denied');
+        }
         
+
         return view('questions.edit',compact('question'));
     }
 
@@ -92,6 +96,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question)){
+            abort(403,'Access denied');
+        }
+        
         $question->delete();
 
         return redirect()->route('questions.index')->with('success',"Your question has been Deleted");
